requirements.txt
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
distlib==0.3.6
distro==1.8.0
filelock==3.12.0
importlib-metadata==6.7.0
numpy==1.21.6
pandas==1.3.5
pdfminer.six==20221105
pdfplumber==0.9.0
Pillow==9.5.0
platformdirs==3.2.0
playsound==1.3.0
pycparser==2.21
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
tabula-py==2.6.0
typing_extensions==4.7.1
virtualenv==20.21.0
Wand==0.6.11
zipp==3.15.0
openpyxl==3.1.2